backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pandas
openpyxl
python-docx
docxtpl
python-multipart
google-generativeai
openai
python-dotenv